# packages/api-contract/openapi-v1.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multicountry Pesticide KG API",
    "version": "1.1"
  },
  "paths": {
...
    "/api/v1/releases": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "Releases",
        "operationId": "releases_api_v1_releases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_ReleaseData__"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/releases/active": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "Active Release",
        "operationId": "active_release_api_v1_releases_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ReleaseData_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/releases/{release_id}": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "Release Detail",
        "operationId": "release_detail_api_v1_releases__release_id__get",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Release Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ReleaseData_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/registration-uses/query": {
      "post": {
        "tags": [
          "registration-uses"
        ],
        "summary": "Registration Uses",
        "operationId": "registration_uses_api_v1_registration_uses_query_post",
        "parameters": [
          {
            "name": "release",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release"
            }
          },
          {
            "name": "X-PestKG-Release",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pestkg-Release"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationUseQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_RegistrationUseData__"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports/registration-uses": {
      "post": {
        "tags": [
          "registration-uses"
        ],
        "summary": "Export Registration Uses",
        "operationId": "export_registration_uses_api_v1_exports_registration_uses_post",
        "parameters": [
          {
            "name": "release",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release"
            }
          },
          {
            "name": "X-PestKG-Release",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pestkg-Release"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_ReleaseData_": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "1.1"
          },
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "data": {
            "$ref": "#/components/schemas/ReleaseData"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "data"
        ],
        "title": "Envelope[ReleaseData]"
      },
...
      "Envelope_list_RegistrationUseData__": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "1.1"
          },
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/RegistrationUseData"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "data"
        ],
        "title": "Envelope[list[RegistrationUseData]]"
      },
      "Envelope_list_ReleaseData__": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "1.1"
          },
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ReleaseData"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "data"
        ],
        "title": "Envelope[list[ReleaseData]]"
      },
      "ErrorDescriptor": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDescriptor"
      },
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "title": "Detail"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDescriptor"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "error"
        ],
        "title": "ErrorEnvelope"
      },
      "ExportRequest": {
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/RegistrationUseFilters"
          }
        },
        "type": "object",
        "title": "ExportRequest"
      },
...
      "RegistrationUseData": {
        "properties": {
          "use_id": {
            "type": "string",
            "title": "Use Id"
          },
          "jurisdiction": {
            "type": "string",
            "title": "Jurisdiction"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "product_label_original": {
            "type": "string",
            "title": "Product Label Original"
          },
          "product_label_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Label En"
          },
          "active_ingredients": {
            "items": {
              "$ref": "#/components/schemas/EntityRef"
            },
            "type": "array",
            "title": "Active Ingredients"
          },
          "crops": {
            "items": {
              "$ref": "#/components/schemas/EntityRef"
            },
            "type": "array",
            "title": "Crops"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/EntityRef"
            },
            "type": "array",
            "title": "Targets"
          },
          "formulations": {
            "items": {
              "$ref": "#/components/schemas/EntityRef"
            },
            "type": "array",
            "title": "Formulations"
          },
          "registration_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Status"
          },
          "registration_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Date"
          },
          "expiry_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          },
          "pairing_status": {
            "type": "string",
            "title": "Pairing Status"
          },
          "source_record_id": {
            "type": "string",
            "title": "Source Record Id"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "use_id",
          "jurisdiction",
          "product_id",
          "product_label_original",
          "product_label_en",
          "active_ingredients",
          "crops",
          "targets",
          "formulations",
          "registration_status",
          "registration_date",
          "expiry_date",
          "pairing_status",
          "source_record_id",
          "source_url"
        ],
        "title": "RegistrationUseData"
      },
      "RegistrationUseFilters": {
        "properties": {
          "jurisdictions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Jurisdictions"
          },
          "query": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "product": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Product"
          },
          "active_ingredient": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Ingredient"
          },
          "crop": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop"
          },
          "target": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "formulation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Formulation"
          },
          "registration_status": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Status"
          },
          "pairing_status": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Pairing Status"
          },
          "valid_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid On"
          }
        },
        "type": "object",
        "title": "RegistrationUseFilters"
      },
      "RegistrationUseQuery": {
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/RegistrationUseFilters"
          },
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "default": 50
          }
        },
        "type": "object",
        "title": "RegistrationUseQuery"
      },
...
      "ReleaseData": {
        "properties": {
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "published_at": {
            "type": "string",
            "title": "Published At"
          },
          "cutoff": {
            "type": "string",
            "title": "Cutoff"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "distribution_status": {
            "type": "string",
            "title": "Distribution Status"
          },
          "known_limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Known Limitations"
          },
          "license": {
            "type": "string",
            "title": "License"
          },
          "inventory": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Inventory"
          },
          "integrity": {
            "$ref": "#/components/schemas/ReleaseIntegrity"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ReleaseArtifact"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "registry_status": {
            "type": "string",
            "title": "Registry Status"
          },
          "registered_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registered At"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "release_id",
          "schema_version",
          "title",
          "published_at",
          "cutoff",
          "status",
          "distribution_status",
          "known_limitations",
          "license",
          "inventory",
          "integrity",
          "artifacts",
          "is_active",
          "registry_status",
          "registered_at"
        ],
        "title": "ReleaseData"
      },
...
    }
  }
}